demo/requirements.txt
gitpython
flask
simplejson
